Footer code samples
