+++ components/footer/src/index.tsx
@@ -1,3 +1,9 @@
+/**
+ * ### References:
+ *
+ * - https://github.com/alphagov/govuk-frontend/blob/main/src/govuk/components/footer/template.njk
+ * - https://github.com/alphagov/govuk-frontend/blob/main/src/govuk/components/footer/_footer.scss
+ */
 import * as React from 'react';
 import styled from 'styled-components';
 import { spacing, typography } from '@govuk-react/lib';
@@ -31,92 +37,10 @@ const FooterContainer = styled('footer')(
 );
 
 /**
+ * The footer provides copyright, licensing and other information about your service and department.
  *
- * ### Usage
- *
- * Simple
- *
- * ```jsx
- * import { Footer } from 'govuk-react'
- *
- * <Footer />
- * ```
- *
- * Footer with Copyright information
- *
- * ```jsx
- * // Import self hosted copyright image
- * import crest from './govuk-crest.png';
- *
- * <Footer
- *   copyright={{
- *     text: 'Crown copyright',
- *     link:
- *       'https://www.nationalarchives.gov.uk/information-management/re-using-public-sector-information/uk-government-licensing-framework/crown-copyright/',
- *     image: {
- *       source: crest,
- *       height: 102,
- *       width: 125,
- *     },
- *   }}
- * />;
- * ```
- *
- * Footer with navigation.  There is also the option to provide footer links with, or without React/Reach Router.
- *
- * ```jsx
- * import { Link } from 'react-router-dom';
- *
- * <Footer>
- *   <Footer.Navigation>
- *     <Footer.NavigationLinks heading="Two column list" listColumns={2}>
- *       <Footer.Link href="/">Navigation item 1</Footer.Link>
- *       <Footer.Link to="/footer-nav-item-2" as={Link}>Navigation item 2 (Router Link)</Footer.Link>
- *       <Footer.Link href="/">Navigation item 3</Footer.Link>
- *       <Footer.Link href="/">Navigation item 4</Footer.Link>
- *       <Footer.Link href="/">Navigation item 5</Footer.Link>
- *       <Footer.Link href="/">Navigation item 6</Footer.Link>
- *     </Footer.NavigationLinks>
- *     <Footer.NavigationLinks heading="Single column list">
- *       <Footer.Link href="/">Navigation item 1</Footer.Link>
- *       <Footer.Link href="/">Navigation item 2</Footer.Link>
- *       <Footer.Link href="/">Navigation item 3</Footer.Link>
- *     </Footer.NavigationLinks>
- *   </Footer.Navigation>
- * </Footer>
- * ```
- *
- * Footer with links in meta area
- *
- * ```jsx
- * <Footer meta={
- *     <Footer.MetaLinks heading="Support links">
- *       <Footer.Link href="/">Item 1</Footer.Link>
- *       <Footer.Link to="/footer-meta-item-2" as={Link}>Item 2 (Router Link)</Footer.Link>
- *       <Footer.Link href="/">Item 3</Footer.Link>
- *     </Footer.MetaLinks>
- *   }
- * />
- * ```
- *
- * Footer with custom content in meta area
- *
- * ```jsx
- * <Footer meta={
- *     <Footer.MetaCustom>
- *       Built by the <Footer.Link href="/">Government Digital Service</Footer.Link>
- *     </Footer.MetaCustom>
- *   }
- * />
- * ```
- *
- * NB <Footer.Link /> is styled extention of @govuk-react/link
- *
- * ### References:
- *
+ * - https://govuk-react.github.io/govuk-react/?path=/docs/footer
  * - https://design-system.service.gov.uk/components/footer/
- * - https://github.com/alphagov/govuk-frontend/blob/main/src/govuk/components/footer/template.njk
- * - https://github.com/alphagov/govuk-frontend/blob/main/src/govuk/components/footer/_footer.scss
  *
  */
 export const Footer: FooterType = ({
